+++ R/gdalcubes_config.R
@@ -1,0 +1,10 @@
+
+
+with_gdalcubes <- function() {
+  if(!requireNamespace("gdalcubes", quietly = TRUE)) {
+    return(invisible(NULL))
+  }
+
+
+
+}