+++ NEWS.md
@@ -7,6 +7,12 @@
   
   - `edl_netrc()` sets up netrc-based access.
   - `edl_unset_netrc()` reverses this, but should rarely be needed.
+  
+  Using other authentication methods in interactive mode will now prompt
+  users to switch to edl_netrc() as the default. 
+  
+* `gdal_cloud_config()` sets recommended settings for GDAL cloud configuration, 
+  see <https://gdalcubes.github.io/source/concepts/config.html#recommended-settings-for-cloud-access>
 
 
 # earthdatalogin 0.0.1